refactor: 보유 금액 입력과 상품 입력 클래스 분리
