+++ src/main/java/vendingmachine/process/ProcessPrepareHoldingAmount.java
@@ -1,0 +1,25 @@
+package vendingmachine.process;
+
+import camp.nextstep.edu.missionutils.Console;
+
+import vendingmachine.user.User;
+import vendingmachine.user.UserGoodsValidation;
+import vendingmachine.user.UserMoneyValidation;
+
+import java.util.List;
+
+public class ProcessPrepareHoldingAmount {
+    public static String holdingAmount;
+
+    public static void inputHoldingAmount() {
+        holdingAmount = User.inputMoney();
+    }
+
+    public static void checkHoldingAmount() {
+        UserMoneyValidation.checkUserMoneyValidation(holdingAmount);
+    }
+
+    public static int toIntegerHoldingAmount() {
+        return Integer.parseInt(holdingAmount);
+    }
+}
